// src/onepunchman/enums/tags.enum.ts
export enum TagsPtBr {}

export enum TagsEn {
  Abilities = '[data-source="abilities"] > div',
  Affiliation = '[data-source="affiliation"] > div',
  Age = "[data-source='age'] > div",
  Alias = '[data-source="alias"] > div',
  Class = '[data-source="class"] > div',
  Description = '#mw-content-text > p',
  Gallery = '#mw-content-text img:not(.sprite)',
  Height = '[data-source="height"] > div',
  Location = '[data-source="location"] > div a',
  Name = '.page-header__title',
  Occupation = '[data-source="occupation"] > div',
  Partner = '[data-source="partners"] > div',
  Photo = '.pi-image-collection > div figure a|.pi-image a',
  Race = '[data-source="race"] > div > a',
  Rank = '[data-source="rank"] > div',
  Sex = '[data-source="gender"] > div',
  Status = '[data-source="status"] > div',
  Weight = '[data-source="weight"] > div'
}

export enum TagsEs {
  Abilities = '[data-source="habilidades"] > div',
  Affiliation = '[data-source="afiliacion"] > div',
  Age = "[data-source='edad'] > div",
  Alias = '[data-source="alias"] > div',
  Class = '[data-source="clase"] > div',
  Description = '#mw-content-text > p',
  Gallery = '#mw-content-text img:not(.sprite)',
  Height = "[data-source='altura'] > div",
  Location = '[data-source="ciudad"] > div a',
  Name = '.page-header__title',
  Occupation = '[data-source="ocupacion"] > div',
  Partner = '[data-source="compañero"] > div',
  Photo = '.pi-image-collection > div figure a|.pi-image a',
  Race = '[data-source="raza"] > div',
  Rank = '[data-source="rango"] > div',
  Sex = '[data-source="genero"] > div',
  Status = '[data-source="estado"] > div',
  Weight = "[data-source='peso'] > div"
}